backend/requirements.txt
fastapi
uvicorn
pydantic
openai-whisper
torch
python-multipart
sqlalchemy
requests 